requirements.txt
plotly==5.1.0
pandas==1.1.5
dash==1.21.0
gunicorn==19.7.1
boto3==1.18.26
python-dotenv==0.19.0
